Fix up test, and add comments.
